refactor: Move warnings to seperate file, use automatic warning type
