+++ src/amltk/sklearn/evaluation.py
@@ -45,6 +45,11 @@ from sklearn.utils.validation import _check_method_params
 
 import amltk.randomness
 from amltk._functional import subclass_map
+from amltk.exceptions import (
+    AutomaticTaskTypeInferredWarning,
+    ImplicitMetricConversionWarning,
+    MismatchedTaskTypeWarning,
+)
 from amltk.optimization.evaluation import EvaluationProtocol
 from amltk.profiling.profiler import Profiler
 from amltk.store import Stored
@@ -82,31 +87,6 @@ _valid_task_types: tuple[TaskTypeName, ...] = (
     "continuous",
     "continuous-multioutput",
 )
-
-
-class ImplicitMetricConversionWarning(UserWarning):
-    """A warning raised when a metric is implicitly converted to an sklearn scorer.
-
-    This is raised when a metric is provided with a custom function and is
-    implicitly converted to an sklearn scorer. This may fail in some cases
-    and it is recommended to explicitly convert the metric to an sklearn
-    scorer with `make_scorer` and then pass it to the metric with
-    [`Metric(fn=...)`][amltk.optimization.Metric].
-    """
-
-
-class TaskTypeWarning(UserWarning):
-    """A warning raised about the task type."""
-
-
-class AutomaticTaskTypeInferredWarning(TaskTypeWarning):
-    """A warning raised when the task type is inferred from the target data."""
-
-
-class MismatchedTaskTypeWarning(TaskTypeWarning):
-    """A warning raised when inferred task type with `task_hint` does not
-    match the inferred task type from the target data.
-    """
 
 
 def _route_params(
